+++ backend/src/main/java/com/EduTech/service/demonstration/DemonstrationService.java
@@ -31,5 +31,6 @@ public interface DemonstrationService {
 		void addDemonstration(DemonstrationFormDTO demonstrationFormDTO,String memId); // 실증 상품 등록 페이지에서 실증 상품 등록하는 기능
 		void updateDemonstration(DemonstrationFormDTO demonstrationFormDTO); // 실증 상품 수정하는 기능
 		void deleteDemonstration(Long demNum); // 실증 번호를 받아 실증 상품을 삭제하는 기능
+		DemonstrationFormDTO selectOne(Long demNum); // 실증 번호를 받아 실증 상품 하나의 정보를 받아오는 기능
 		
 }
