+++ client/src/pages/demonstration/UpdateDemPage.js
@@ -1,8 +1,13 @@
+import DemTitleComponent from "../../components/demonstration/DemTitleComponent";
+import "react-datepicker/dist/react-datepicker.css";
+import UpdateComponent from "../../components/demonstration/UpdateComponent";
+import { useParams } from "react-router-dom";
 const UpdateDemPage = () => {
+    const {demNum}=useParams();
     return (
-        <div>
-            <h1>실증 상품 업데이트 페이지</h1>
-            <br></br>
+        <div className="w-full">
+            <DemTitleComponent title="실증 수정" />
+            <UpdateComponent demNum={demNum}/>
         </div>
     );
 }
